+++ spec/hatchet/python_update_warning_spec.rb
@@ -79,38 +79,25 @@ RSpec.describe 'Python update warnings' do
   end
 
   context 'with a runtime.txt containing python-3.9.0' do
-    let(:allow_failure) { false }
-    let(:app) { Hatchet::Runner.new('spec/fixtures/python_3.9_outdated', allow_failure:) }
+    let(:app) { Hatchet::Runner.new('spec/fixtures/python_3.9_outdated') }
 
-    context 'when using Heroku-22 or older', stacks: %w[heroku-20 heroku-22] do
-      include_examples 'warns there is a Python update available', '3.9.0', LATEST_PYTHON_3_9
-    end
-
-    context 'when using Heroku-24', stacks: %w[heroku-24] do
-      let(:allow_failure) { true }
-
-      # We only support Python 3.9 on Heroku-22 and older.
-      include_examples 'aborts the build without showing an update warning', '3.9.0'
-    end
+    include_examples 'warns there is a Python update available', '3.9.0', LATEST_PYTHON_3_9
   end
 
   context 'with a runtime.txt containing python-3.10.0' do
-    let(:allow_failure) { false }
-    let(:app) { Hatchet::Runner.new('spec/fixtures/python_3.10_outdated', allow_failure:) }
+    let(:app) { Hatchet::Runner.new('spec/fixtures/python_3.10_outdated') }
 
     include_examples 'warns there is a Python update available', '3.10.0', LATEST_PYTHON_3_10
   end
 
   context 'with a runtime.txt containing python-3.11.0' do
-    let(:allow_failure) { false }
-    let(:app) { Hatchet::Runner.new('spec/fixtures/python_3.11_outdated', allow_failure:) }
+    let(:app) { Hatchet::Runner.new('spec/fixtures/python_3.11_outdated') }
 
     include_examples 'warns there is a Python update available', '3.11.0', LATEST_PYTHON_3_11
   end
 
   context 'with a runtime.txt containing python-3.12.0' do
-    let(:allow_failure) { false }
-    let(:app) { Hatchet::Runner.new('spec/fixtures/python_3.12_outdated', allow_failure:) }
+    let(:app) { Hatchet::Runner.new('spec/fixtures/python_3.12_outdated') }
 
     include_examples 'warns there is a Python update available', '3.12.0', LATEST_PYTHON_3_12
   end